pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "brandfield-reporting"
version = "0.1.0"
description = "Automated reporting system for BrandField agency"
requires-python = ">=3.11"
dependencies = [
    "facebook-business>=19.0",
    "jinja2>=3.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "responses>=0.25",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
brandfield = "brandfield.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["brandfield*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
